+++ src/Drawer.js
@@ -4,10 +4,12 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import ContainerRender from 'rc-util/lib/ContainerRender';
 import getScrollBarSize from 'rc-util/lib/getScrollBarSize';
-import { dataToArray, transitionEnd } from './utils';
+import { dataToArray, transitionEnd, addEventListener, removeEventListener, transformArguments } from './utils';
 
 const IS_REACT_16 = 'createPortal' in ReactDOM;
 
+const currentDrawer = {};
+let bodyAddTransitionEnd = false;
 const windowIsUndefined = typeof window === 'undefined';
 class Drawer extends React.PureComponent {
   static defaultProps = {
@@ -15,12 +17,16 @@ class Drawer extends React.PureComponent {
     placement: 'left',
     getContainer: 'body',
     level: 'all',
-    levelTransition: 'transform .3s cubic-bezier(0.78, 0.14, 0.15, 0.86)',
-    onChange: () => {},
-    onMaskClick: () => {},
-    onHandleClick: () => {},
-    handleChild: <i className="drawer-handle-icon" />,
-    handleStyle: {},
+    duration: '.3s',
+    ease: 'cubic-bezier(0.78, 0.14, 0.15, 0.86)',
+    onChange: () => { },
+    onMaskClick: () => { },
+    onHandleClick: () => { },
+    handler: (
+      <div className="drawer-handle">
+        <i className="drawer-handle-icon" />
+      </div>
+    ),
     showMask: true,
     maskStyle: {},
     wrapperClassName: '',
@@ -32,23 +38,51 @@ class Drawer extends React.PureComponent {
     this.levelDom = [];
     this.contextDom = null;
     this.maskDom = null;
-    this.handleDom = null;
+    this.handlerdom = null;
     this.mousePos = null;
     this.firstEnter = false;// 记录首次进入.
-    if (props.onIconClick || props.parent || props.iconChild || props.width) {  // eslint-disable-line react/prop-types
-      /* eslint-disable  no-console */
-      console.warn(
+    this.timeout = null;
+    this.drawerId = Number((Date.now() + Math.random()).toString()
+      .replace('.', Math.round(Math.random() * 9))).toString(16);
+
+    if (props.onIconClick || props.parent || props.iconChild || props.width // eslint-disable-line react/prop-types
+      || props.handleChild) {  // eslint-disable-line react/prop-types
+      console.warn(// eslint-disable-line no-console
         'rc-drawer-menu API has been changed, please look at the releases, ' +
         'https://github.com/react-component/drawer-menu/releases'
       );
     }
+    const open = props.open !== undefined ? props.open : !!props.defaultOpen;
+    currentDrawer[this.drawerId] = open;
     this.state = {
-      open: props.open !== undefined ? props.open : !!props.defaultOpen,
+      open,
     };
   }
   componentDidMount() {
+    if (!windowIsUndefined) {
+      if (!bodyAddTransitionEnd) {
+        addEventListener(
+          document.body,
+          transitionEnd,
+          this.onBodyTransitionEnd
+        );
+        bodyAddTransitionEnd = true;
+      }
+      let passiveSupported = false;
+      window.addEventListener(
+        'test',
+        null,
+        Object.defineProperty({}, 'passive', {
+          get: () => {
+            passiveSupported = true;
+            return null;
+          },
+        })
+      );
+      this.passive = passiveSupported ? { passive: false } : false;
+    }
     const open = this.getOpen();
-    if (this.props.handleChild || open) {
+    if (this.props.handler || open) {
       this.getDefault(this.props);
       this.forceUpdate();
     }
@@ -91,13 +125,22 @@ class Drawer extends React.PureComponent {
         this.container.parentNode.removeChild(this.container);
       }
     }
+    if (!windowIsUndefined) {
+      removeEventListener(
+        document.body,
+        transitionEnd,
+        this.onBodyTransitionEnd
+      );
+    }
+    clearTimeout(this.timeout);
+    delete currentDrawer[this.drawerId];
     // suppport react15
     if (IS_REACT_16) {
       return;
     }
     this.renderComponent({
       afterClose: this.removeContainer,
-      onClose() {},
+      onClose() { },
       visible: false,
     });
   }
@@ -121,6 +164,24 @@ class Drawer extends React.PureComponent {
       open: !open,
     });
   };
+
+  onBodyTransitionEnd = (e) => {
+    if (e.target === e.currentTarget) {
+      document.body.style.transform = '';
+      document.body.style.transition = '';
+      document.body.style.position = '';
+      if (this.dom) {
+        this.dom.style.transition = '';
+      }
+      if (!this.state.open) {
+        document.body.style.overflow = '';
+        if (this.props.placement === 'right' && this.maskDom) {
+          this.maskDom.style.right = '';
+          this.maskDom.style.width = '';
+        }
+      }
+    }
+  }
 
   getDefault = props => {
     this.getParentAndLevelDom(props);
@@ -170,37 +231,54 @@ class Drawer extends React.PureComponent {
   };
 
   setLevelDomTransform = (open, openTransition, placementName, value) => {
-    const { placement, levelTransition, onChange } = this.props;
-    this.levelDom.forEach(dom => {
-      if (this.isOpenChange || openTransition) {
-        /* eslint no-param-reassign: "error" */
-        dom.style.transition = levelTransition;
-        dom.addEventListener(transitionEnd, this.trnasitionEnd);
-      }
-      const placementPos = placement === 'left' || placement === 'top' ? value : -value;
-      dom.style.transform = open ? `${placementName}(${placementPos}px)` : '';
-    });
-    // 处理 body 滚动
+    const { placement, levelMove, duration, ease, onChange } = this.props;
     if (!windowIsUndefined) {
+      this.levelDom.forEach(dom => {
+        if (this.isOpenChange || openTransition) {
+          /* eslint no-param-reassign: "error" */
+          dom.style.transition = `transform ${duration} ${ease}`;
+          addEventListener(dom, transitionEnd, this.trnasitionEnd);
+          let levelValue = open ? value : 0;
+          if (levelMove) {
+            const $levelMove = transformArguments(levelMove, { target: dom, open });
+            levelValue = open ? $levelMove[0] : $levelMove[1] || 0;
+          }
+          const placementPos = placement === 'left' || placement === 'top' ? levelValue : -levelValue;
+          dom.style.transform = placementPos ? `${placementName}(${placementPos}px)` : '';
+        }
+      });
+      // 处理 body 滚动
       const eventArray = ['touchstart'];
-      const domArray = [document.body, this.maskDom, this.handleDom, this.contextDom];
-      let passiveSupported = false;
-      window.addEventListener(
-        'test',
-        null,
-        Object.defineProperty({}, 'passive', {
-          get: () => {
-            passiveSupported = true;
-            return null;
-          },
-        })
-      );
-      const passive = passiveSupported ? { passive: false } : false;
-      if (open) {
+      const domArray = [document.body, this.maskDom, this.handlerdom, this.contextDom];
+      const right = getScrollBarSize();
+      const widthTransition = `width ${duration} ${ease}`;
+      const trannsformTransition = `transform ${duration} ${ease}`;
+      if (open && document.body.style.overflow !== 'hidden') {
         document.body.style.overflow = 'hidden';
-        const right = getScrollBarSize();
         if (right) {
-          document.body.style.paddingRight = `${right}px`;
+          document.body.style.position = 'relative';
+          document.body.style.width = `calc(100% - ${right}px)`;
+          document.body.style.transition = 'none';
+          this.dom.style.transition = 'none';
+          switch (placement) {
+            case 'right':
+              this.dom.style.transform = `translateX(-${right}px)`;
+              break;
+            case 'top':
+            case 'bottom':
+              this.dom.style.width = `calc(100% - ${right}px)`;
+              break;
+            default:
+              break;
+          }
+          clearTimeout(this.timeout);
+          this.timeout = setTimeout(() => {
+            document.body.style.transition = widthTransition;
+            this.dom.style.transition = `${trannsformTransition},${widthTransition}`;
+            document.body.style.width = '';
+            this.dom.style.width = '';
+            this.dom.style.transform = '';
+          });
         }
         // 手机禁滚
         if (document.body.addEventListener) {
@@ -211,13 +289,42 @@ class Drawer extends React.PureComponent {
             item.addEventListener(
               eventArray[i] || 'touchmove',
               i ? this.removeMoveHandler : this.removeStartHandler,
-              passive
+              this.passive
             );
           });
         }
-      } else {
+      } else if (!Object.keys(currentDrawer).some(key => currentDrawer[key])) {
         document.body.style.overflow = '';
-        document.body.style.paddingRight = '';
+        if (this.isOpenChange && right) {
+          document.body.style.position = 'relative';
+          document.body.style.overflowX = 'hidden'
+          document.body.style.transition = 'none';
+          document.body.style.width = `calc(100% + ${right}px)`;
+          this.dom.style.transition = 'none';
+          switch (placement) {
+            case 'right': {
+              this.dom.style.transform = `translateX(${right}px)`;
+              this.maskDom.style.right = `${right}px`;
+              this.maskDom.style.width = `calc(100% + ${right}px)`;
+              break;
+            }
+            case 'top':
+            case 'bottom': {
+              this.dom.style.width = `calc(100% + ${right}px)`;
+              break;
+            }
+            default:
+              break;
+          }
+          clearTimeout(this.timeout);
+          this.timeout = setTimeout(() => {
+            document.body.style.transition = widthTransition;
+            this.dom.style.transition = `${trannsformTransition},${widthTransition}`;
+            document.body.style.width = '';
+            this.dom.style.transform = '';
+            this.dom.style.width = '';
+          });
+        }
         if (document.body.removeEventListener) {
           domArray.forEach((item, i) => {
             if (!item) {
@@ -226,12 +333,13 @@ class Drawer extends React.PureComponent {
             item.removeEventListener(
               eventArray[i] || 'touchmove',
               i ? this.removeMoveHandler : this.removeStartHandler,
-              passive
+              this.passive
             );
           });
         }
       }
     }
+
     if (onChange && this.isOpenChange && this.firstEnter) {
       onChange(open);
       this.isOpenChange = false;
@@ -245,8 +353,7 @@ class Drawer extends React.PureComponent {
       style,
       placement,
       children,
-      handleChild,
-      handleStyle,
+      handler,
       showMask,
       maskStyle,
     } = this.props;
@@ -258,8 +365,7 @@ class Drawer extends React.PureComponent {
     const value = this.contextDom
       ? this.contextDom.getBoundingClientRect()[
         placement === 'left' || placement === 'right' ? 'width' : 'height'
-      ]
-      : 0;
+      ] : 0;
     const placementName = `translate${placement === 'left' || placement === 'right' ? 'X' : 'Y'}`;
     // 百分比与像素动画不同步，第一次打用后全用像素动画。
     const defaultValue = !this.contextDom ? '100%' : `${value}px`;
@@ -269,10 +375,22 @@ class Drawer extends React.PureComponent {
     if (this.isOpenChange === undefined || this.isOpenChange) {
       this.setLevelDomTransform(open, false, placementName, value);
     }
+    const handlerCildren = handler && React.cloneElement(handler, {
+      onClick: (e) => {
+        if (handler.props.onClick) {
+          handler.props.onClick();
+        }
+        this.onIconTouchEnd(e);
+      },
+      ref: (c) => {
+        this.handlerdom = c;
+      }
+    });
     return (
       <div
         className={wrapperClassname}
         style={style}
+        ref={c => { this.dom = c; }}
       >
         {showMask && (
           <div
@@ -284,7 +402,10 @@ class Drawer extends React.PureComponent {
             }}
           />
         )}
-        <div className={`${prefixCls}-content-wrapper`} style={{ transform }}>
+        <div
+          className={`${prefixCls}-content-wrapper`}
+          style={{ transform }}
+        >
           <div
             className={`${prefixCls}-content`}
             ref={c => {
@@ -295,18 +416,7 @@ class Drawer extends React.PureComponent {
           >
             {children}
           </div>
-          {handleChild && (
-            <div
-              className={`${prefixCls}-handle`}
-              onClick={this.onIconTouchEnd}
-              style={handleStyle}
-              ref={c => {
-                this.handleDom = c;
-              }}
-            >
-              {handleChild}
-            </div>
-          )}
+          {handlerCildren}
         </div>
       </div>
     );
@@ -335,7 +445,7 @@ class Drawer extends React.PureComponent {
     const differY = e.changedTouches[0].clientY - this.startPos.y;
     if (
       currentTarget === this.maskDom ||
-      currentTarget === this.handleDom ||
+      currentTarget === this.handlerdom ||
       (currentTarget === this.contextDom &&
         ((((currentTarget.scrollTop + currentTarget.offsetHeight >= currentTarget.scrollHeight &&
           differY < 0) ||
@@ -351,7 +461,7 @@ class Drawer extends React.PureComponent {
   };
 
   trnasitionEnd = e => {
-    e.target.removeEventListener(transitionEnd, this.trnasitionEnd);
+    removeEventListener(e.target, transitionEnd, this.trnasitionEnd);
     e.target.style.transition = '';
   };
 
@@ -370,6 +480,7 @@ class Drawer extends React.PureComponent {
   render() {
     const { getContainer, wrapperClassName } = this.props;
     const open = this.getOpen();
+    currentDrawer[this.drawerId] = open ? this.container : open;
     const children = this.getChildToRender(this.firstEnter ? open : false);
     if (!getContainer) {
       return (
@@ -419,10 +530,11 @@ Drawer.propTypes = {
   prefixCls: PropTypes.string,
   placement: PropTypes.string,
   level: PropTypes.oneOfType([PropTypes.string, PropTypes.array]),
-  levelTransition: PropTypes.string,
+  levelMove: PropTypes.oneOfType([PropTypes.number, PropTypes.func, PropTypes.array]),
+  ease: PropTypes.string,
+  duration: PropTypes.string,
   getContainer: PropTypes.oneOfType([PropTypes.string, PropTypes.func, PropTypes.object, PropTypes.bool]),
-  handleChild: PropTypes.any,
-  handleStyle: PropTypes.object,
+  handler: PropTypes.any,
   onChange: PropTypes.func,
   onMaskClick: PropTypes.func,
   onHandleClick: PropTypes.func,